pyproject.toml
[tool.poetry]
name = "ayelet_initmanager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pika = "^1.2.0"
aio-pika = "^8.0.3"
fastapi = "^0.78.0"
aioredis = "^1.3.1"
aiohttp = "^3.8.1"
uvicorn = "^0.17.6"
msgpack = "^1.0.3"
rabbitmq = "^0.2.0"
orjson = "^3.6.8"
ormsgpack = "^1.2.2"
azure-storage-blob = "~12.11.0"
envolved = "^0.5.0"
retrying = "^1.3.3"
async_asgi_testclient = "^1.4.11"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-asyncio = "0.18.3"
isort="^5.10.1"
flake8 = {version = "*", allow-prereleases = true}
mypy = "^0.961"
coverage = "^6.3.2"
yellowbox = {version="^0.7.1", extras=["azure_storage", "redis", "rabbit_mq"]}



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
